GenBackCounts.cpp works. Modified GenCondorScriptsHist.py. Need to check that it works.

+++ BackgroundRecon/GenCondorScriptsHist.py
@@ -13,10 +13,10 @@ if __name__=="__main__":
     parser.add_argument("output_path", help="Absolute path to output directory. Folder will be created if not present.")
     parser.add_argument("Process_name", help="Name assigned to .cmd and .sh files, the default output of condor, and output histogram file.")
     parser.add_argument("NBatches", type=int, help="Number of reconstruction files under consideration")
-    parser.add_argument("-w","--weighted", action='store_true', help="Flag to turn on reweighing scheme. If on, need to provide the total number of events, total exposure time, as well asRefFlux,PhysicalFlux, and EffectiveArea root files")
     parser.add_argument("--RefFlux", default="", help=" Name of Root file (not the path) containing 1/E reference flux (Log uniform)")
     parser.add_argument("--PhysicalFlux", default="", help="Name of Root file (not the path) containing physical flux of gamma ray source")
     parser.add_argument("--EffectiveArea", default="", help="Name of Effective Area root file (not the path)")
+    parser.add_argument("--MaskFile", default="", help="Name of Mask root file (not the path)")
     parser.add_argument("--TotalEvents", default=-1, type=int, help="Total number of events generated across the entire condor job")
     parser.add_argument("--ExposureTime", default=-1.0,type=float, help="Total exposure time in seconds")
     parser.add_argument("-mem","--MemoryRequested", default=300, type=int, help="Amount of memory in MB requested per job")
@@ -29,15 +29,15 @@ if __name__=="__main__":
     condor_error_name = args.Process_name+"_$(Process).err"
     condor_log_name = args.Process_name+"_$(Process).log"
     tar_name = args.Process_name+".tar.gz"
-    # Create absolute path to GenSkyMap for histograms, shell script path, Reconstructed data location (ie. input), and output file name
+    # Create absolute path to GenBackCounts for histograms, shell script path, Reconstructed data location (ie. input), and output file name
     home = os.getcwd()
     shell_path = os.path.join(home,shell_script_name)
-    GenSkyMap_path = os.path.join(home,"SkyMap","GenSkyMap")
+    GenSkyMap_path = os.path.join(home,"BackgroundCount")
     tar_path = os.path.join(home,tar_name)
     ReconstructedDataFileName = args.Input_Path_Rootname+"$(Process).root"
     ShellReconstructedDataFileName = args.Input_Path_Rootname+"${process}.root"
     ReconstructionData_path = os.path.join(args.Input_Path,ReconstructedDataFileName)
-    OutputFileName = args.Process_name+"_SkyMap_${process}.root"
+    OutputFileName = args.Process_name+"_BackCount_${process}.root"
     ## Check if directory exists. If it does not, create the folder
     if not (os.path.exists(args.output_path)):
         try:
@@ -47,27 +47,30 @@ if __name__=="__main__":
             sys.exit()
   ## Check that required root files are present in current directory 
   ## Also check if Total Events and Exposure time are valid
-    if(args.weighted):
-        if((args.RefFlux=="")  or (args.PhysicalFlux=="") or (args.EffectiveArea=="")):
-            print("Need to input Reference Flux, PhysicalFlux, and Effective Area names when doing reweighting")
-            sys.exit()
-        RefFluxPath = os.path.join(home,"SkyMap",args.RefFlux)
-        PhysFluxPath = os.path.join(home,"SkyMap",args.PhysicalFlux)
-        EffAreaPath = os.path.join(home,"SkyMap",args.EffectiveArea)
-        if (not os.path.exists(RefFluxPath)):
-            print("Reference Flux file not present in SkyMap")
-            sys.exit()
-        if (not os.path.exists(PhysFluxPath)):
-            print("Physical Flux file not present in SkyMap")
-            sys.exit()
-        if (not os.path.exists(EffAreaPath)):
-            print("Effective Area file not present in SkyMap")
-            sys.exit()
-        if (args.TotalEvents <=0):
-            print("Number of events needs to be larger than zero")
-            sys.exit()
-        if (args.ExposureTime <=0):
-            print("Exposure time needs to be larger than zero")
+    if((args.RefFlux=="")  or (args.PhysicalFlux=="") or (args.EffectiveArea=="") or (args.MaskFile=="")):
+        print("Need to input Reference Flux, PhysicalFlux, Effective Area and Mask File names for reweighting")
+        sys.exit()
+    RefFluxPath = os.path.join(home,"BackgroundCount",args.RefFlux)
+    PhysFluxPath = os.path.join(home,"BackgroundCount",args.PhysicalFlux)
+    EffAreaPath = os.path.join(home,"BackgroundCount",args.EffectiveArea)
+    MaskPath = os.path.join(home,"BackgroundCount",args.MaskFile)
+    if (not os.path.exists(RefFluxPath)):
+        print("Reference Flux file not present in SkyMap")
+        sys.exit()
+    if (not os.path.exists(PhysFluxPath)):
+        print("Physical Flux file not present in SkyMap")
+        sys.exit()
+    if (not os.path.exists(EffAreaPath)):
+        print("Effective Area file not present in SkyMap")
+        sys.exit()
+    if (not os.path.exists(MaskPath)):
+        print("Effective Area file not present in SkyMap")
+        sys.exit()
+    if (args.TotalEvents <=0):
+        print("Number of events needs to be larger than zero")
+        sys.exit()
+    if (args.ExposureTime <=0):
+        print("Exposure time needs to be larger than zero")
     ## Write the shell script
     with open(shell_script_name,'w') as f:
         f.write("#!/bin/bash\n")
@@ -78,12 +81,9 @@ if __name__=="__main__":
         f.write("tar -xzf "+tar_name+"\n")
         f.write("mv "+ShellReconstructedDataFileName+" SkyMap\n")
         f.write("cd SkyMap\n")
-        if(args.weighted):
-            f.write("./GenSkyMap -i "+ShellReconstructedDataFileName+" -o "+OutputFileName)
-            f.write(" -t "+str(args.ExposureTime)+" --TotalEvents " +str(args.TotalEvents))
-            f.write(" --EffAreaFile "+args.EffectiveArea+" --PhysicalFluxFile "+args.PhysicalFlux+" --ReferenceFluxFile "+args.RefFlux+"\n")
-        else:
-            f.write("./GenSkyMap -i "+ShellReconstructedDataFileName+" -o "+OutputFileName+"\n")
+        f.write("./GenBackCounts -i "+ShellReconstructedDataFileName+" -o "+OutputFileName)
+        f.write(" -t "+str(args.ExposureTime)+" --TotalEvents " +str(args.TotalEvents))
+        f.write(" --EffAreaFile "+args.EffectiveArea+" --PhysicalFluxFile "+args.PhysicalFlux+" --ReferenceFluxFile "+args.RefFlux+" --MaskFile "+ args.MaskFile+"\n")
         f.write("mv "+OutputFileName+" ..\n")
         f.write("cd ..\n")
     ## Write the cmd file
@@ -103,4 +103,4 @@ if __name__=="__main__":
         f.write("log = "+condor_log_name+"\n")
         f.write("notification   = Never\n")
         f.write("queue "+str(args.NBatches)+"\n")
-    subprocess.run(["tar", "-czf", tar_name, "SkyMap/"]) 
+    subprocess.run(["tar", "-czf", tar_name, "BackgroundCount/"])
